move KTX2 parsing to a webworker, un-monkeypatch msc_basis_transcoder.js (built off commit 78790b5)
